+++ src/__tests__/screen-reader-only-test.tsx
@@ -1,22 +1,22 @@
 import * as React from 'react';
-import {render} from '@testing-library/react';
+import {render, screen} from '@testing-library/react';
 import ScreenReaderOnly from '../screen-reader-only';
 
 test('Renders no wrapper if it has one single child', () => {
-    const {asFragment} = render(
-        <ScreenReaderOnly>
+    render(
+        <ScreenReaderOnly dataAttributes={{testid: 'test'}}>
             <span>Some text</span>
         </ScreenReaderOnly>
     );
-    expect(asFragment().firstChild?.nodeName).toBe('SPAN');
+    expect(screen.getByTestId('test').nodeName).toBe('SPAN');
 });
 
 test('Renders a wrapper if it has several children', () => {
-    const {asFragment} = render(
-        <ScreenReaderOnly>
+    render(
+        <ScreenReaderOnly dataAttributes={{testid: 'test'}}>
             <span>Some text</span>
             <p>Other text</p>
         </ScreenReaderOnly>
     );
-    expect(asFragment().firstChild?.nodeName).toBe('DIV');
+    expect(screen.getByTestId('test').nodeName).toBe('DIV');
 });