+++ src/skins/blau.tsx
@@ -176,10 +176,7 @@ export const getBlauSkin: GetKnownSkin = () => {
             tagBackgroundSuccess: palette.blauGreen10,
             tagBackgroundWarning: palette.blauYellow10,
             tagBackgroundError: palette.blauRed10,
-            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(palette.grey6, 0)} 0%, ${applyAlpha(
-                palette.grey6,
-                0.4
-            )} 30%, ${applyAlpha(palette.grey6, 0.7)} 100%)`,
+            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(palette.grey6, 0)} 0%, ${applyAlpha(palette.grey6, 0.4)} 30%, ${applyAlpha(palette.grey6, 0.7)} 100%)`,
         },
         darkModeColors: {
             background: palette.darkModeBlack,
@@ -311,10 +308,7 @@ export const getBlauSkin: GetKnownSkin = () => {
             tagBackgroundSuccess: palette.darkModeGrey6,
             tagBackgroundWarning: palette.darkModeGrey6,
             tagBackgroundError: palette.darkModeGrey6,
-            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(palette.grey6, 0)} 0%, ${applyAlpha(
-                palette.grey6,
-                0.4
-            )} 30%, ${applyAlpha(palette.grey6, 0.7)} 100%)`,
+            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(palette.grey6, 0)} 0%, ${applyAlpha(palette.grey6, 0.4)} 30%, ${applyAlpha(palette.grey6, 0.7)} 100%)`,
         },
         borderRadii: {
             avatar: '50%',