+++ src/skins/o2-new.tsx
@@ -184,10 +184,7 @@ export const getO2NewSkin: GetKnownSkin = () => {
             tagBackgroundSuccess: palette.o2Green10,
             tagBackgroundWarning: palette.o2Orange10,
             tagBackgroundError: palette.o2Red10,
-            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(
-                palette.beyondBlue,
-                0
-            )} 0%, ${applyAlpha(palette.beyondBlue, 0.4)} 30%, ${applyAlpha(palette.beyondBlue, 1)} 100%)`,
+            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(palette.beyondBlue, 0)} 0%, ${applyAlpha(palette.beyondBlue, 0.4)} 30%, ${applyAlpha(palette.beyondBlue, 1)} 100%)`,
         },
         darkModeColors: {
             background: palette.darkModeBlack,
@@ -319,10 +316,7 @@ export const getO2NewSkin: GetKnownSkin = () => {
             tagBackgroundSuccess: palette.darkModeGrey6,
             tagBackgroundWarning: palette.darkModeGrey6,
             tagBackgroundError: palette.darkModeGrey6,
-            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(palette.black, 0)} 0%, ${applyAlpha(
-                palette.black,
-                0.4
-            )} 30%, ${applyAlpha(palette.black, 0.7)} 100%)`,
+            cardContentOverlay: `linear-gradient(180deg, ${applyAlpha(palette.black, 0)} 0%, ${applyAlpha(palette.black, 0.4)} 30%, ${applyAlpha(palette.black, 0.7)} 100%)`,
         },
         borderRadii: {
             avatar: '50%',