+++ src/theme.tsx
@@ -68,6 +68,7 @@ const TEXTS_ES = {
     timerSecondsLongLabel: 'segundos',
     timerDisplayMinutesLabel: 'min',
     timerDisplaySecondsLabel: 'seg',
+    tableActionsHeaderLabel: 'Acciones',
 };
 
 const TEXTS_EN: ThemeTexts = {
@@ -131,6 +132,7 @@ const TEXTS_EN: ThemeTexts = {
     timerSecondsLongLabel: 'seconds',
     timerDisplayMinutesLabel: 'min',
     timerDisplaySecondsLabel: 'sec',
+    tableActionsHeaderLabel: 'Actions',
 };
 
 const TEXTS_DE: ThemeTexts = {
@@ -194,6 +196,7 @@ const TEXTS_DE: ThemeTexts = {
     timerSecondsLongLabel: 'Sekunden',
     timerDisplayMinutesLabel: 'Min.',
     timerDisplaySecondsLabel: 'Sek.',
+    tableActionsHeaderLabel: 'Aktionen',
 };
 
 const TEXTS_PT: ThemeTexts = {
@@ -257,6 +260,7 @@ const TEXTS_PT: ThemeTexts = {
     timerSecondsLongLabel: 'segundos',
     timerDisplayMinutesLabel: 'min',
     timerDisplaySecondsLabel: 'seg',
+    tableActionsHeaderLabel: 'Ações',
 };
 
 export const getTexts = (locale: Locale): typeof TEXTS_ES => {
@@ -332,28 +336,25 @@ const getReactRouter5Link = (ReactRouterLink: React.ComponentType<any>): LinkCom
 
 const getReactRouter6Link =
     (ReactRouterLink: React.ComponentType<any>): LinkComponent =>
-    ({innerRef, ...props}) =>
-        <ReactRouterLink ref={innerRef} {...props} />;
+    ({innerRef, ...props}) => <ReactRouterLink ref={innerRef} {...props} />;
 
 const getNext12Link =
     (NextLink: React.ComponentType<any>): LinkComponent =>
-    ({to, innerRef, children, ...props}) =>
-        (
-            <NextLink href={to}>
-                <a ref={innerRef} {...props}>
-                    {children}
-                </a>
-            </NextLink>
-        );
+    ({to, innerRef, children, ...props}) => (
+        <NextLink href={to}>
+            <a ref={innerRef} {...props}>
+                {children}
+            </a>
+        </NextLink>
+    );
 
 const getNext13Or14Link =
     (NextLink: React.ComponentType<any>): LinkComponent =>
-    ({to, innerRef, children, ...props}) =>
-        (
-            <NextLink href={to} ref={innerRef} {...props}>
-                {children}
-            </NextLink>
-        );
+    ({to, innerRef, children, ...props}) => (
+        <NextLink href={to} ref={innerRef} {...props}>
+            {children}
+        </NextLink>
+    );
 
 export const getMisticaLinkComponent = (Link?: ThemeConfig['Link']): LinkComponent => {
     if (!Link) {